lower the number of DB connections
- share a single MongoClient for each service
- close the MongoClients when the app is destroyed
