+++ _backend/services/user-service/src/main/java/com/accounts/MongoUtil.java
@@ -2,25 +2,26 @@ package com.accounts;
 
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
-import jakarta.annotation.PostConstruct;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.inject.Disposes;
+import jakarta.enterprise.inject.Produces;
 
 @ApplicationScoped
 public class MongoUtil {
 
-    private static MongoClient client;
-
-    @PostConstruct
-    public void init() {
+    @Produces
+    @ApplicationScoped
+    public MongoClient createMongoClient() {
         System.out.println(">>>> Creating Mongo Connection");
-        client = MongoClients.create(getConnectionString());
+        return MongoClients.create(getConnectionString());
+    }
+
+    public void closeMongoClient(@Disposes MongoClient client) {
+        System.out.println(">>>> Closing Mongo Connection");
+        client.close();
     }
 
     private static String getConnectionString() {
         return System.getenv("CONNECTION_STRING");
-    }
-
-    public MongoClient getMongoClient(){
-        return client;
     }
 }
