+++ _backend/services/user-service/src/main/java/com/notifications/NotificationResource.java
@@ -43,7 +43,11 @@ import static com.mongodb.client.model.Filters.eq;
 @Path("/notifications")
 public class NotificationResource {
 
+    @Inject
     NotificationService notificationService;
+
+    @Inject
+    AccountService accountService;
 
     @GET
     @Path("/user/{userId}")
@@ -57,7 +61,6 @@ public class NotificationResource {
     @Operation(summary = "Get all notifications for a specific user",
             description = "Returns JSON of all notifications where the user is the recipient, enter ID of user recieving notifications")
     public Response getNotificationsForUser(@PathParam("userId") String userId, @Context HttpHeaders headers) {
-        NotificationService notificationService = new NotificationService();
         String authHeader = headers.getHeaderString(HttpHeaders.AUTHORIZATION);
 
         if (authHeader == null || !authHeader.toLowerCase().startsWith("bearer ")) {
@@ -68,7 +71,7 @@ public class NotificationResource {
 
         String jwtString = authHeader.replaceFirst("(?i)^Bearer\\s+", "");
 
-        Document userDoc = notificationService.accountService.retrieveUserFromJWT(jwtString);
+        Document userDoc = accountService.retrieveUserFromJWT(jwtString);
 
         if (userDoc == null) {
             return Response.status(Response.Status.UNAUTHORIZED).entity(new Document("error", "User not authorized to have notifications").toJson()).build();
@@ -131,7 +134,6 @@ public class NotificationResource {
             description = "notification data required: from(id), to(id), type(string), and quote_id(id)"
     )
     public Response createNotification(String jsonInput, @Context HttpHeaders headers) {
-        NotificationService notificationService = new NotificationService();
         String authHeader = headers.getHeaderString(HttpHeaders.AUTHORIZATION);
 
         if (authHeader == null || !authHeader.toLowerCase().startsWith("bearer ")) {
@@ -142,7 +144,7 @@ public class NotificationResource {
 
         String jwtString = authHeader.replaceFirst("(?i)^Bearer\\s+", "");
 
-        Document userDoc = notificationService.accountService.retrieveUserFromJWT(jwtString);
+        Document userDoc = accountService.retrieveUserFromJWT(jwtString);
 
         if (userDoc == null) {
             return Response.status(Response.Status.UNAUTHORIZED).entity(new Document("error", "User not authorized to create notification").toJson()).build();
@@ -207,11 +209,11 @@ public class NotificationResource {
 
             toUser.remove("expires_at");
 
-            Account toAccount = notificationService.accountService.document_to_account(toUser);
+            Account toAccount = accountService.document_to_account(toUser);
 
             toAccount.Notifications.add(notificationId.toString());
 
-            notificationService.accountService.updateUser(toAccount.toJson(), toId);
+            accountService.updateUser(toAccount.toJson(), toId);
 
             JsonObject response = Json.createObjectBuilder()
                     .add("success", true)
@@ -243,7 +245,6 @@ public class NotificationResource {
     @Operation(summary = "Delete a notification by ID",
             description = "Deletes a notification with the specified ID")
     public Response deleteNotification(@PathParam("notificationId") String notificationId, @Context HttpHeaders headers) {
-        NotificationService notificationService = new NotificationService();
         String authHeader = headers.getHeaderString(HttpHeaders.AUTHORIZATION);
 
         if (authHeader == null || !authHeader.toLowerCase().startsWith("bearer ")) {
@@ -254,7 +255,7 @@ public class NotificationResource {
 
         String jwtString = authHeader.replaceFirst("(?i)^Bearer\\s+", "");
 
-        Document userDoc = notificationService.accountService.retrieveUserFromJWT(jwtString);
+        Document userDoc = accountService.retrieveUserFromJWT(jwtString);
 
         if (userDoc == null) {
             return Response.status(Response.Status.UNAUTHORIZED).entity(new Document("error", "User not authorized to delete notifications").toJson()).build();
@@ -303,7 +304,6 @@ public class NotificationResource {
     @Operation(summary = "Retrieve a notification by ID",
             description = "Retrieves a notification by notification ID and returns the necessary info based on the type of notification")
     public Response getNotification(@PathParam("notificationId") String notificationId, @Context HttpHeaders headers) {
-        NotificationService notificationService = new NotificationService();
         String authHeader = headers.getHeaderString(HttpHeaders.AUTHORIZATION);
 
         if (authHeader == null || !authHeader.toLowerCase().startsWith("bearer ")) {
@@ -314,7 +314,7 @@ public class NotificationResource {
 
         String jwtString = authHeader.replaceFirst("(?i)^Bearer\\s+", "");
 
-        Document userDoc = notificationService.accountService.retrieveUserFromJWT(jwtString);
+        Document userDoc = accountService.retrieveUserFromJWT(jwtString);
 
         if (userDoc == null) {
             return Response.status(Response.Status.UNAUTHORIZED).entity(new Document("error", "User not authorized to retrieve notifications").toJson()).build();
